% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{AIC2}
\alias{AIC2}
\title{Akaike Information Criterion (AIC)}
\usage{
AIC2(model)
}
\description{
Calculates the Akaike Information Criterion (AIC) for a model.
AIC = -2\emph{log-likelihood + 2}k, where k is the number of parameters.
Lower AIC values indicate better models.
}
