% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{AICc}
\alias{AICc}
\title{Corrected Akaike Information Criterion (AICc)}
\usage{
AICc(model)
}
\description{
Calculates the corrected Akaike Information Criterion (AICc) for small samples.
AICc = AIC + 2\emph{k}(k+1)/(n-k-1). This correction reduces bias when the
sample size is small relative to the number of parameters.
}
