% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesLaw.R
\name{bayesLaw}
\alias{bayesLaw}
\title{Bayes Law Calculator}
\usage{
bayesLaw(conmat = NULL, pwild = NULL, sensitivity = NULL, specificity = NULL)
}
\arguments{
\item{conmat}{Confusion matrix (2x2) for the test}

\item{pwild}{Incidence in the population}

\item{sensitivity}{True positive rate for the test}

\item{specificity}{True negative rate for the test}
}
\value{
The function returns the probability that the patient is diseased, given the disease prevalence in the population and either the confusion matrix or the sensitivity and specificity.
}
\description{
\code{bayesLaw} performs basic calculations using Bayes Law
}
\details{
This function calculates the posterior probability of disease given a positive test result
(positive predictive value) using Bayes' theorem. The calculation requires:

\deqn{P(Disease|Positive) = \frac{P(Positive|Disease) \times P(Disease)}{P(Positive)}}

where \eqn{P(Positive)} can be expanded as:
\deqn{P(Positive) = P(Positive|Disease) \times P(Disease) + P(Positive|No Disease) \times P(No Disease)}

The function can accept input in two different formats:
\enumerate{
\item A 2x2 confusion matrix along with population incidence
\item Sensitivity and specificity values along with population incidence
}

When using a confusion matrix, the function will extract sensitivity and specificity
from the matrix. The confusion matrix should be structured as:
\preformatted{
          Predicted
Actual    Positive Negative
  Positive TP      FN
  Negative FP      TN
}

Note: All probabilities should be provided as values between 0 and 1.
}
\examples{
# Using sensitivity and specificity
bayesLaw(sensitivity = 0.95, specificity = 0.90, pwild = 0.01)

# Using a confusion matrix
conf_matrix <- matrix(c(95, 5, 10, 90), nrow = 2, byrow = TRUE,
                     dimnames = list(c("Diseased", "Healthy"),
                                     c("Test+", "Test-")))
bayesLaw(conmat = conf_matrix, pwild = 0.01)

}
\seealso{
For more information on Bayes' theorem: \url{https://en.wikipedia.org/wiki/Bayes\%27_theorem}
}
