% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPermutations.R
\name{getPermutations}
\alias{getPermutations}
\title{Permutation Generator}
\usage{
getPermutations(n)
}
\arguments{
\item{n}{An integer}
}
\value{
An n! x n matrix containing all permutations
}
\description{
This produces \emph{all} permutations of the numbers 1 through \code{n}.
}
\details{
This function generates all permutations of the integers from 1 to \code{n} using a recursive algorithm.
The algorithm works by:
\enumerate{
\item Generating all permutations of size \code{n-1}
\item For each permutation, inserting the value \code{n} in all possible positions
\item Adjusting values greater than or equal to the inserted value by incrementing them by 1
}

The time complexity is O(n!) and the space complexity is O(n!) as well, since all permutations are stored in memory.
}
\examples{
# All permutations of 1, 2, 3
permutations(3)

# All permutations of 1, 2, 3, 4
perm4 <- permutations(4)
dim(perm4)  # 24 rows, 4 columns

# First few permutations of 4 elements
head(permutations(4))

# Use permutations to shuffle data
set.seed(123)
data <- letters[1:4]
perm <- permutations(4)
data[perm[5, ]]  # Apply the 5th permutation to the data

}
\references{
Knuth, D. E. (1997). \emph{The Art of Computer Programming, Volume 1 (3rd ed.): Fundamental Algorithms}. Addison-Wesley. pp. 1-2.

This algorithm is similar to the "Algorithm L (Lexicographic permutation generation)" described by Knuth, though implemented recursively rather than iteratively.
}
\seealso{
\code{\link[utils:combn]{utils::combn()}} for combinations, \code{\link[=sample]{sample()}} for random permutations
}
