% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{QAIC}
\alias{QAIC}
\title{Quasi-Akaike Information Criterion (QAIC)}
\usage{
QAIC(model)
}
\description{
Calculates the Quasi-Akaike Information Criterion (QAIC) for models with
overdispersion, such as some generalized linear models. QAIC adjusts AIC
for extra variation not captured by the model.
}
