% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{QAICc}
\alias{QAICc}
\title{Corrected Quasi-Akaike Information Criterion (QAICc)}
\usage{
QAICc(model)
}
\description{
Calculates the corrected Quasi-Akaike Information Criterion (QAICc) for
small samples with overdispersion.
}
