% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rice.R
\docType{data}
\name{rice}
\alias{rice}
\title{Yields of four varieties of rice}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{variety}{variety of the rice grown in the plot}
\item{yield}{yield (in bushels per acre) of the rice in the plot}
}
}
\source{
Generic data
}
\usage{
rice
}
\description{
A dataset containing rice yields (four varieties) planted in 16 plots:
}
\keyword{datasets}
