% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{SBC}
\alias{SBC}
\title{Schwarz Bayesian Criterion (SBC/BIC)}
\usage{
SBC(model)
}
\description{
Calculates the Schwarz Bayesian Criterion (SBC), also known as the
Bayesian Information Criterion (BIC). SBC = -2*log-likelihood + log(n)*k,
where n is sample size and k is number of parameters. This criterion
penalizes complex models more heavily than AIC, especially with larger samples.
}
