% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFit.R
\name{SBCc}
\alias{SBCc}
\title{Corrected Schwarz Bayesian Criterion (SBCc)}
\usage{
SBCc(model)
}
\description{
Calculates the corrected Schwarz Bayesian Criterion (SBCc) for small samples.
This is a small-sample correction to SBC similar to the AICc correction.
}
